Fixed some bugs and delete some unused stuff
