+++ Project/app/src/main/java/org/ionproject/android/common/dto/ICalendarDto.kt
@@ -8,7 +8,6 @@ import com.fasterxml.jackson.annotation.JsonProperty
 data class ICalendarDto(
     @JsonProperty("class") val clazz: List<String>? = null,
     val properties: CalendarProperties,
-    val entities: List<SubEntity>,
     val actions: List<SirenAction>,
     val links: List<SirenLink>
 )
@@ -23,3 +22,48 @@ data class CalendarCreator(
     val prodid: Value,
     val version: Value
 )
+
+data class EventProperties(
+    val type: String? = null,
+    val properties: ComponentProperties
+)
+
+data class ComponentProperties(
+    val uid: Value,
+    val summary: Value? = null,
+    val description: Value? = null,
+    val categories: Value? = null,
+    val created: Value? = null,
+    val dtstamp: Value? = null,
+    val dtstart: Value? = null,
+    val dtend: Value? = null,
+    val duration: Value? = null,
+    val rrule: Rules? = null,
+    val attachment: Value? = null,
+    val due: Value? = null,
+    val relatedTo: List<Related>? = null
+)
+
+data class Value(
+    val value: List<String>
+)
+
+data class Rules(
+    val value: RuleValues
+)
+
+data class RuleValues(
+    val freq: String,
+    val until: String,
+    val byDay: List<String>
+)
+
+data class Related(
+    val parameters: RelType? = null,
+    val value: String
+)
+
+data class RelType(
+    val reltype: String
+)
+
