--- Project/app/src/main/java/org/ionproject/android/common/dto/JournalDto.kt
@@ -1,15 +1,0 @@
-package org.ionproject.android.common.dto
-
-import com.fasterxml.jackson.annotation.JsonProperty
-
-data class JournalDto(
-    @JsonProperty("class") val clazz: List<String>? = null,
-    val properties: ComponentProperties,
-    val entities: List<SubEntity>,
-    val links: List<Link>
-)
-
-data class Link(
-    val rel: List<String>,
-    val href: List<String>
-)
