--- Project/app/src/main/java/org/ionproject/android/common/dto/TodoDto.kt
@@ -1,10 +1,0 @@
-package org.ionproject.android.common.dto
-
-import com.fasterxml.jackson.annotation.JsonProperty
-
-data class TodoDto(
-    @JsonProperty("class") val clazz: List<String>? = null,
-    val properties: ComponentProperties,
-    val entities: List<SubEntity>,
-    val links: List<SirenLink>
-)
