--- Project/app/src/main/java/org/ionproject/android/common/dto/enums/EventType.kt
@@ -1,8 +1,0 @@
-package org.ionproject.android.common.dto.enums
-
-enum class EventType {
-    Exam,
-    Todo,
-    Lecture,
-    Journal
-}
