+++ Project/app/src/main/java/org/ionproject/android/common/ionwebapi/MockIonWebAPI.kt
@@ -206,8 +206,7 @@ private val e2ndExam =
 
 // Calendar
 private const val pg11DCalendar =
-    "{\"class\":[\"calendar\"],\"properties\":{\"type\":\"calendar\",\"properties\":{\"prodid\":{\"value\":\"/v0/courses/1/classes/1920v/sections/11D\"},\"version\":{\"value\":\"2.0\"}},\"subComponents\":[{\"type\":\"event\",\"properties\":{\"uid\":{\"value\":\"event/45678\"},\"summary\":{\"value\":\"Theory Class PG-1920v\"},\"description\":{\"value\":\"Theory Class of the PG-1920v class.\"},\"categories\":{\"value\":[\"Lecture\",\"Theory Class\"]},\"dtstamp\":{\"value\":\"2020-02-10T10:34:24Z\"},\"dtstart\":{\"value\":\"2020-03-19T11:00:00Z\"},\"duration\":{\"value\":\"PT03H00M00S\"},\"rrule\":{\"value\":{\"freq\":\"WEEKLY\",\"until\":\"2020-06-10T00:00:00Z\",\"byDay\":[\"TH\"]}}}},{\"type\":\"event\",\"properties\":{\"uid\":{\"value\":\"event/45679\"},\"summary\":{\"value\":\"Theory Class PG-1920v\"},\"description\":{\"value\":\"Theory Class of the PG-1920v class.\"},\"categories\":{\"value\":[\"Lecture\",\"Theory Class\"]},\"dtstamp\":{\"value\":\"2020-02-10T10:34:24Z\"},\"dtstart\":{\"value\":\"2020-03-16T12:30:00Z\"},\"duration\":{\"value\":\"PT01H30M00S\"},\"rrule\":{\"value\":{\"freq\":\"WEEKLY\",\"until\":\"2020-06-10T00:00:00Z\",\"byDay\":[\"MO\"]}}}},{\"type\":\"event\",\"properties\":{\"uid\":{\"value\":\"event/1234\"},\"summary\":{\"value\":\"PG 1st Exam\"},\"description\":{\"value\":\"First exam of the PG course during the 1920v semester\"},\"categories\":{\"value\":\"Exam\"},\"created\":{\"value\":\"2020-02-10T10:34:20Z\"},\"dtstamp\":{\"value\":\"2020-02-10T10:34:20Z\"},\"dtstart\":{\"value\":\"2020-06-19T14:00:00Z\"},\"dtend\":{\"value\":\"2020-06-19T16:30:00Z\"}}},{\"type\":\"event\",\"properties\":{\"uid\":{\"value\":\"event/1235\"},\"summary\":{\"value\":\"PG 2nd Exam\"},\"description\":{\"value\":\"Second exam of the PG course during the 1920v semester\"},\"categories\":{\"value\":\"Exam\"},\"created\":{\"value\":\"2020-02-10T10:34:20Z\"},\"dtstamp\":{\"value\":\"2020-02-10T10:34:20Z\"},\"dtstart\":{\"value\":\"2020-07-11T14:00:00Z\"},\"dtend\":{\"value\":\"2020-07-11T16:30:00Z\"}}},{\"type\":\"todo\",\"properties\":{\"uid\":{\"value\":\"todo/123490\"},\"summary\":{\"value\":\"PG 1st Serie\"},\"description\":{\"value\":\"First serie of exercises for the PG course during the 1920v semester\"},\"attachment\":{\"value\":\"https://github.com/isel-leic-pg/1920v-public/wiki/phase-1\"},\"categories\":{\"value\":[\"Evaluation\",\"Assignment\",\"Programming\"]},\"created\":{\"value\":\"2020-02-10T10:34:20Z\"},\"dtstamp\":{\"value\":\"2020-02-12T12:24:50Z\"},\"due\":{\"value\":\"2020-06-08T23:59:59Z\"}}},{\"type\":\"journal\",\"properties\":{\"uid\":{\"value\":\"journal/123497\"},\"summary\":{\"value\":\"Error in 1st PG Exam\"},\"description\":{\"value\":\"The first exercise had and ambigous question that resulted in both answers B and C being correct. Therefore both will be awarded full marks.\"},\"categories\":{\"value\":[\"Error\",\"Evaluation\",\"Exam\",\"Programming\"]},\"created\":{\"value\":\"2020-02-10T10:34:20Z\"},\"dtstamp\":{\"value\":\"2020-02-12T12:24:50Z\"},\"dtstart\":{\"value\":\"2020-03-19T14:00:00Z\"},\"relatedTo\":[{\"value\":\"/v0/courses/1/classes/1920v/calendar/1234\"},{\"parameters\":{\"reltype\":\"SIBLING\"},\"value\":\"/v0/courses/1/classes/1920v/calendar/123485\"}]}}]},\"entities\":[{\"class\":[\"class\",\"section\"],\"rel\":[\"/rel/class-section\"],\"properties\":{\"uid\":\"61D\",\"lecturer\":1010},\"links\":[{\"rel\":[\"self\"],\"href\":\"/v0/courses/1/classes/1920v\"}]}],\"actions\":[{\"name\":\"search\",\"summary\":\"Search components\",\"method\":\"GET\",\"href\":\"/v0/courses/1/classes/1920v/calendar{?type,startBefore,startAfter,endBefore,endAfter,summary}\",\"isTemplated\":true,\"type\":\"application/x-www-form-urlencoded\",\"fields\":[{\"name\":\"type\",\"type\":\"text\",\"class\":\"https://example.org/param/free-text-query\"},{\"name\":\"startBefore\",\"type\":\"date\",\"class\":\"https://example.org/param/date-query\"},{\"name\":\"startAfter\",\"type\":\"date\",\"class\":\"https://example.org/param/date-query\"},{\"name\":\"endBefore\",\"type\":\"date\",\"class\":\"https://example.org/param/date-query\"},{\"name\":\"endAfter\",\"type\":\"date\",\"class\":\"https://example.org/param/date-query\"},{\"name\":\"summary\",\"type\":\"text\",\"class\":\"https://example.org/param/free-text-query\"}]},{\"name\":\"add-item\",\"title\":\"Add Item\",\"method\":\"POST\",\"href\":\"/v0/courses/1/classes/1920v/calendar\",\"isTemplated\":false,\"type\":\"application/json\",\"fields\":[]},{\"name\":\"batch-delete\",\"title\":\"Delete multiple items\",\"method\":\"DELETE\",\"isTemplated\":true,\"href\":\"/v0/courses/1/classes/1920v/calendar{?type}\",\"fields\":[{\"name\":\"type\",\"type\":\"text\",\"class\":\"https://example.org/param/free-text-query\"}]}],\"links\":[{\"rel\":[\"self\"],\"href\":\"/v0/courses/1/classes/1920v/61D/calendar\"},{\"rel\":[\"about\"],\"href\":\"/v0/courses/1/classes/1920v/61D\"}]}"
-
+    "{\"class\":[\"calendar\"],\"properties\":{\"type\":\"calendar\",\"properties\":{\"prodid\":{\"value\":\"/v0/courses/1/classes/1920v/sections/11D\"},\"version\":{\"value\":\"2.0\"}},\"subComponents\":[{\"type\":\"event\",\"properties\":{\"uid\":{\"value\":\"event/45678\"},\"summary\":{\"value\":\"Theory Class PG-1920v\"},\"description\":{\"value\":\"Theory Class of the PG-1920v class.\"},\"categories\":{\"value\":[\"Lecture\",\"Theory Class\"]},\"dtstamp\":{\"value\":\"2020-02-10T10:34:24Z\"},\"dtstart\":{\"value\":\"2020-03-19T11:00:00Z\"},\"duration\":{\"value\":\"PT03H00M00S\"},\"rrule\":{\"value\":{\"freq\":\"WEEKLY\",\"until\":\"2020-06-10T00:00:00Z\",\"byDay\":[\"TH\"]}}}},{\"type\":\"event\",\"properties\":{\"uid\":{\"value\":\"event/45679\"},\"summary\":{\"value\":\"Theory Class PG-1920v\"},\"description\":{\"value\":\"Theory Class of the PG-1920v class.\"},\"categories\":{\"value\":[\"Lecture\",\"Theory Class\"]},\"dtstamp\":{\"value\":\"2020-02-10T10:34:24Z\"},\"dtstart\":{\"value\":\"2020-03-16T12:30:00Z\"},\"duration\":{\"value\":\"PT01H30M00S\"},\"rrule\":{\"value\":{\"freq\":\"WEEKLY\",\"until\":\"2020-06-10T00:00:00Z\",\"byDay\":[\"MO\"]}}}},{\"type\":\"event\",\"properties\":{\"uid\":{\"value\":\"event/1234\"},\"summary\":{\"value\":\"PG 1st Exam\"},\"description\":{\"value\":\"First exam of the PG course during the 1920v semester\"},\"categories\":{\"value\":\"Exam\"},\"created\":{\"value\":\"2020-02-10T10:34:20Z\"},\"dtstamp\":{\"value\":\"2020-02-10T10:34:20Z\"},\"dtstart\":{\"value\":\"2020-06-19T14:00:00Z\"},\"dtend\":{\"value\":\"2020-06-19T16:30:00Z\"}}},{\"type\":\"event\",\"properties\":{\"uid\":{\"value\":\"event/1235\"},\"summary\":{\"value\":\"PG 2nd Exam\"},\"description\":{\"value\":\"Second exam of the PG course during the 1920v semester\"},\"categories\":{\"value\":\"Exam\"},\"created\":{\"value\":\"2020-02-10T10:34:20Z\"},\"dtstamp\":{\"value\":\"2020-02-10T10:34:20Z\"},\"dtstart\":{\"value\":\"2020-07-11T14:00:00Z\"},\"dtend\":{\"value\":\"2020-07-11T16:30:00Z\"}}},{\"type\":\"todo\",\"properties\":{\"uid\":{\"value\":\"todo/123490\"},\"summary\":{\"value\":\"PG 1st Serie\"},\"description\":{\"value\":\"First serie of exercises for the PG course during the 1920v semester\"},\"attachment\":{\"value\":\"https://github.com/isel-leic-pg/1920v-public/wiki/phase-1\"},\"categories\":{\"value\":[\"Evaluation\",\"Assignment\",\"Programming\"]},\"created\":{\"value\":\"2020-02-10T10:34:20Z\"},\"dtstamp\":{\"value\":\"2020-02-12T12:24:50Z\"},\"due\":{\"value\":\"2020-06-08T23:59:59Z\"}}},{\"type\":\"journal\",\"properties\":{\"uid\":{\"value\":\"journal/123497\"},\"summary\":{\"value\":\"Error in 1st PG Exam\"},\"description\":{\"value\":\"The first exercise had and ambigous question that resulted in both answers B and C being correct. Therefore both will be awarded full marks.\"},\"categories\":{\"value\":[\"Error\",\"Evaluation\",\"Exam\",\"Programming\"]},\"created\":{\"value\":\"2020-02-10T10:34:20Z\"},\"dtstamp\":{\"value\":\"2020-02-12T12:24:50Z\"},\"dtstart\":{\"value\":\"2020-03-19T14:00:00Z\"},\"relatedTo\":[{\"value\":\"/v0/courses/1/classes/1920v/calendar/1234\"},{\"parameters\":{\"reltype\":\"SIBLING\"},\"value\":\"/v0/courses/1/classes/1920v/calendar/123485\"}]}}]},\"actions\":[{\"name\":\"search\",\"summary\":\"Search components\",\"method\":\"GET\",\"href\":\"/v0/courses/1/classes/1920v/calendar{?type,startBefore,startAfter,endBefore,endAfter,summary}\",\"isTemplated\":true,\"type\":\"application/x-www-form-urlencoded\",\"fields\":[{\"name\":\"type\",\"type\":\"text\",\"class\":\"https://example.org/param/free-text-query\"},{\"name\":\"startBefore\",\"type\":\"date\",\"class\":\"https://example.org/param/date-query\"},{\"name\":\"startAfter\",\"type\":\"date\",\"class\":\"https://example.org/param/date-query\"},{\"name\":\"endBefore\",\"type\":\"date\",\"class\":\"https://example.org/param/date-query\"},{\"name\":\"endAfter\",\"type\":\"date\",\"class\":\"https://example.org/param/date-query\"},{\"name\":\"summary\",\"type\":\"text\",\"class\":\"https://example.org/param/free-text-query\"}]},{\"name\":\"add-item\",\"title\":\"Add Item\",\"method\":\"POST\",\"href\":\"/v0/courses/1/classes/1920v/calendar\",\"isTemplated\":false,\"type\":\"application/json\",\"fields\":[]},{\"name\":\"batch-delete\",\"title\":\"Delete multiple items\",\"method\":\"DELETE\",\"isTemplated\":true,\"href\":\"/v0/courses/1/classes/1920v/calendar{?type}\",\"fields\":[{\"name\":\"type\",\"type\":\"text\",\"class\":\"https://example.org/param/free-text-query\"}]}],\"links\":[{\"rel\":[\"self\"],\"href\":\"/v0/courses/1/classes/1920v/61D/calendar\"},{\"rel\":[\"about\"],\"href\":\"/v0/courses/1/classes/1920v/61D\"}]}"
 // Lectures mock
 private val pg11Dlectures =
     "{\"class\":[\"calendar\"],\"properties\":{\"type\":\"calendar\",\"properties\":{\"prodid\":{\"value\":\"/v0/courses/1/classes/1920v/sections/11D\"},\"version\":{\"value\":\"2.0\"}},\"subComponents\":[{\"type\":\"event\",\"properties\":{\"uid\":{\"value\":\"event/45678\"},\"summary\":{\"value\":\"Theory Class WAD-1920v\"},\"description\":{\"value\":\"Theory Class of the WAD-1920v class.\"},\"categories\":{\"value\":[\"Lecture\",\"Theory Class\"]},\"dtstamp\":{\"value\":\"2020-02-10T10:34:24Z\"},\"dtstart\":{\"value\":\"2020-03-19T11:00:00Z\"},\"duration\":{\"value\":\"PT03H00M00S\"},\"rrule\":{\"value\":{\"freq\":\"WEEKLY\",\"until\":\"2020-06-10T00:00:00Z\",\"byDay\":[\"TH\"]}}}},{\"type\":\"event\",\"properties\":{\"uid\":{\"value\":\"event/45679\"},\"summary\":{\"value\":\"Theory Class WAD-1920v\"},\"description\":{\"value\":\"Theory Class of the WAD-1920v class.\"},\"categories\":{\"value\":[\"Lecture\",\"Theory Class\"]},\"dtstamp\":{\"value\":\"2020-02-10T10:34:24Z\"},\"dtstart\":{\"value\":\"2020-03-16T11:00:00Z\"},\"duration\":{\"value\":\"PT01H30M00S\"},\"rrule\":{\"value\":{\"freq\":\"WEEKLY\",\"until\":\"2020-06-10T00:00:00Z\",\"byDay\":[\"MO\"]}}}}]},\"entities\":[{\"class\":[\"class\",\"section\"],\"rel\":[\"/rel/class-section\"],\"properties\":{\"uid\":\"61D\",\"lecturer\":1010},\"links\":[{\"rel\":[\"self\"],\"href\":\"/v0/courses/1/classes/1920v\"}]}],\"actions\":[{\"name\":\"search\",\"summary\":\"Search components\",\"method\":\"GET\",\"href\":\"/v0/courses/1/classes/1920v/calendar{?type,startBefore,startAfter,endBefore,endAfter,summary}\",\"isTemplated\":true,\"type\":\"application/x-www-form-urlencoded\",\"fields\":[{\"name\":\"type\",\"type\":\"text\",\"class\":\"https://example.org/param/free-text-query\"},{\"name\":\"startBefore\",\"type\":\"date\",\"class\":\"https://example.org/param/date-query\"},{\"name\":\"startAfter\",\"type\":\"date\",\"class\":\"https://example.org/param/date-query\"},{\"name\":\"endBefore\",\"type\":\"date\",\"class\":\"https://example.org/param/date-query\"},{\"name\":\"endAfter\",\"type\":\"date\",\"class\":\"https://example.org/param/date-query\"},{\"name\":\"summary\",\"type\":\"text\",\"class\":\"https://example.org/param/free-text-query\"}]},{\"name\":\"add-item\",\"title\":\"Add Item\",\"method\":\"POST\",\"href\":\"/v0/courses/1/classes/1920v/calendar\",\"isTemplated\":false,\"type\":\"application/json\",\"fields\":[]},{\"name\":\"batch-delete\",\"title\":\"Delete multiple items\",\"method\":\"DELETE\",\"isTemplated\":true,\"href\":\"/v0/courses/1/classes/1920v/calendar{?type}\",\"fields\":[{\"name\":\"type\",\"type\":\"text\",\"class\":\"https://example.org/param/free-text-query\"}]}],\"links\":[{\"rel\":[\"self\"],\"href\":\"/v0/courses/1/classes/1920v/61D/calendar\"},{\"rel\":[\"about\"],\"href\":\"/v0/courses/1/classes/1920v/61D\"}]}"
